Move SES tests, convert to vitest
- Move SES tests from `javascriptv3/example_code/tests`
to `javascriptv3/example_code/ses/tests`.
- Convert SES tests from Jest to vitest.
- Fix license issue.
- Fix typo.
