+++ javascriptv3/example_code/ses/tests/get-template.integration.test.js
@@ -1,5 +1,7 @@
-import { run, TEMPLATE_NAME } from "../../ses/src/ses_gettemplate";
-import { createTemplate, deleteTemplate } from "../../ses/src/libs/sesUtils";
+import { describe, beforeAll, afterAll, it, expect } from "vitest";
+
+import { run, TEMPLATE_NAME } from "../src/ses_gettemplate";
+import { createTemplate, deleteTemplate } from "../src/libs/sesUtils";
 
 describe("ses_gettemplate", () => {
   beforeAll(async () => {
