+++ javascriptv3/example_code/ses/tests/list-identities.integration.test.js
@@ -1,6 +1,8 @@
+import { describe, beforeAll, afterAll, it, expect } from "vitest";
+
 import { getUniqueName, postfix } from "../../libs/utils/util-string.js";
-import { createIdentity, deleteIdentity } from "../../ses/src/libs/sesUtils.js";
-import { run } from "../../ses/src/ses_listidentities";
+import { createIdentity, deleteIdentity } from "../src/libs/sesUtils.js";
+import { run } from "../src/ses_listidentities";
 
 describe("ses_listidentities", () => {
   const IDENTITY_NAME = postfix(getUniqueName("IdentityName"), "@example.com");
