--- javascriptv3/example_code/tests/ses/babel.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-    presets: ['@babel/preset-env']
-}
