--- javascriptv3/example_code/tests/ses/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "aws-sdk-v3-iam-examples",
-  "version": "1.0.0",
-  "main": "index.js",
-  "repository": "git@github.com/awsdocs/aws-doc-sdk-examples/tree/master/javascriptv3/example_code/iam.git",
-  "author": "Brian Murray <brmur@amazon.com>, Alex Forsyth <alex-git@amazon.com>",
-  "license": "Apache 2.0",
-  "dependencies": {
-    "@aws-sdk/client-s3": "^3.150.0",
-    "@aws-sdk/client-ses": "^3.32.0",
-    "@aws-sdk/node-http-handler": "^3.32.0",
-    "@aws-sdk/types": "^3.32.0",
-    "jest": "^26.6.3",
-    "node-fetch": "^2.6"
-  },
-  "scripts": {
-    "test": "jest"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.18.10",
-    "@babel/preset-env": "^7.18.10",
-    "@types/jest": "^28.1.6",
-    "@types/node": "^14.0.23",
-    "typescript": "^4.0.2"
-  }
-}
